server/requirements.txt
pip==21.0.1
fastapi==0.63.0
psycopg2-binary==2.8.6
uvicorn[standard]==0.13.3
loguru==0.5.3
aiofiles==0.6.0
Jinja2==2.11.3
